+++ projects/angular-split/src/public_api.ts
@@ -5,5 +5,4 @@
 export { AngularSplitModule } from './lib/module'
 export { SplitComponent } from './lib/component/split.component'
 export { SplitAreaDirective } from './lib/directive/split-area.directive'
-export { IOutputData, IOutputAreaSizes } from './lib/interface'
-export { ANGULAR_SPLIT_DEFAULT_OPTIONS } from './lib/angular-split-config.token'
+export * from './lib/interface'